pyproject.toml
[tool.poetry]
name = "drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
mypy = "^1.13.0"
pillow = "^11.0.0"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
ipython = "^8.30.0"
coverage = "^7.6.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
